+++ example/lib/pages/default_example.dart
@@ -57,7 +57,7 @@ class _DefaultExampleState extends State<DefaultExample>
                       configs: const ProImageEditorConfigs(),
                       theme: ThemeData.dark(),
                     );
-                    var url = 'https://picsum.photos/2000';
+                    var url = 'https://picsum.photos/5000';
                     var bytes = await fetchImageAsUint8List(url);
 
                     if (!context.mounted) return;