+++ lib/utils/decode_image.dart
@@ -29,33 +29,48 @@ Future<ImageInfos> decodeImageInfos({
   double heightRatio = (rotated ? w : h).toDouble() / screenSize.height;
   double pixelRatio = max(heightRatio, widthRatio);
 
+  Size renderedSize =
+      Size(w.toDouble() / pixelRatio, h.toDouble() / pixelRatio);
+
   return ImageInfos(
     rawSize: Size(w.toDouble(), h.toDouble()),
-    renderedSize: Size(w.toDouble() / pixelRatio, h.toDouble() / pixelRatio),
+    renderedSize: renderedSize,
+    cropRectSize: configs != null && configs.isNotEmpty
+        ? configs.cropRect.size
+        : renderedSize,
     pixelRatio: pixelRatio,
+    isRotated: configs?.is90DegRotated ?? false,
   );
 }
 
 class ImageInfos {
   final Size rawSize;
   final Size renderedSize;
+  final Size cropRectSize;
   final double pixelRatio;
+  final bool isRotated;
 
   const ImageInfos({
     required this.rawSize,
     required this.renderedSize,
+    required this.cropRectSize,
     required this.pixelRatio,
+    required this.isRotated,
   });
 
   ImageInfos copyWith({
     Size? rawSize,
     Size? renderedSize,
+    Size? cropRectSize,
     double? pixelRatio,
+    bool? isRotated,
   }) {
     return ImageInfos(
       rawSize: rawSize ?? this.rawSize,
       renderedSize: renderedSize ?? this.renderedSize,
+      cropRectSize: cropRectSize ?? this.cropRectSize,
       pixelRatio: pixelRatio ?? this.pixelRatio,
+      isRotated: isRotated ?? this.isRotated,
     );
   }
 }